+++ lib/lenna/middleware/app.rb
@@ -1,12 +1,11 @@
 # frozen_string_literal: true
 
+require 'singleton'
+
 module Lenna
   module Middleware
     # The MiddlewareManager class is responsible for managing the middlewares.
     #
-    # @attr mutex                   [Mutex] the mutex used to synchronize the
-    #                                       access to the global middlewares and
-    #                                       the middleware chains cache.
     # @attr global_middlewares      [Array] the global middlewares
     # @attr middleware_chains_cache [Hash]  the middleware chains cache
     #
@@ -14,15 +13,19 @@ module Lenna
     #       The middlewares that are added to a specific route are added to the
     #       global middlewares.
     #
-    # @api private
-    #
     # @since 0.1.0
+    #
     class App
-      # @return [Mutex] the mutex used to synchronize the access to the global
-      attr_reader :global_middlewares
+      include Singleton
 
-      # @return [Hash] the middleware chains cache
-      attr_reader :middleware_chains_cache
+      # @return [Mutex] the mutex used to synchronize the access to the global
+      #                 middlewares.
+      #
+      attr_accessor :global_middlewares
+      # @return [Mutex] the mutex used to synchronize the access to the
+      #                 middleware chains cache.
+      #
+      attr_accessor :middleware_chains_cache
 
       # This method will initialize the global middlewares and the
       # middleware chains cache.
@@ -30,8 +33,22 @@ module Lenna
       # @return [void]
       #
       # @since 0.1.0
+      #
       def initialize
-        @mutex = ::Mutex.new
+        @global_middlewares      = []
+        @middleware_chains_cache = {}
+      end
+
+      # This method is used to reset the global middlewares and the middleware
+      # chains cache.
+      #
+      # @return [void]
+      #
+      # @see #initialize
+      #
+      # @api public
+      #
+      def reset!
         @global_middlewares      = []
         @middleware_chains_cache = {}
       end
@@ -41,11 +58,10 @@ module Lenna
       # @return            [void]
       #
       # @since 0.1.0
+      #
       def use(middlewares)
-        @mutex.synchronize do
-          @global_middlewares += Array(middlewares)
-          @middleware_chains_cache = {}
-        end
+        @global_middlewares += Array(middlewares)
+        @middleware_chains_cache = {}
       end
 
       # This method is used to fetch or build the middleware chain for the given
@@ -58,13 +74,22 @@ module Lenna
       # @see #build_middleware_chain
       #
       # @since 0.1.0
-      def fetch_or_build_middleware_chain(action, route_middlewares)
-        middleware_signature = action.object_id.to_s
+      #
+      def fetch_or_build_middleware_chain(
+        action,
+        route_middlewares,
+        http_method: nil,
+        path: nil
+      )
+        signature =
+          if http_method && path
+            [http_method, path, route_middlewares].hash.to_s
+          else
+            ['global', route_middlewares].hash.to_s
+          end
 
-        @mutex.synchronize do
-          @middleware_chains_cache[middleware_signature] ||=
-            build_middleware_chain(action, route_middlewares)
-        end
+        @middleware_chains_cache[signature] ||=
+          build_middleware_chain(action, route_middlewares)
       end
 
       # This method is used to build the middleware chain for the given action
@@ -83,8 +108,9 @@ module Lenna
       #                            mw1 -> mw2 -> action
       #                            The action will be the last middleware in the
       #                            chain.
+      #
       def build_middleware_chain(action, middlewares)
-        all_middlewares = @global_middlewares + Array(middlewares)
+        all_middlewares = (@global_middlewares + Array(middlewares)).uniq
 
         all_middlewares.reverse.reduce(action) do |next_middleware, middleware|
           ->(req, res) {