+++ lib/lenna/router/request.rb
@@ -7,7 +7,24 @@ module Lenna
     # @attr headers [Hash] the request headers
     # @attr body    [Hash] the request body
     # @attr params  [Hash] the request params
+    #
     class Request < ::Rack::Request
+      # This method is used to set the request headers.
+      #
+      # @param name  [String] the header name
+      # @param value [String] the header value
+      #
+      # @return [String] the header value
+      #
+      # @api public
+      #
+      # @example:
+      #   request.put_header('Foo', 'bar')
+      #
+      #  request.headers
+      #  # => { 'Foo' => 'bar' }
+      #
+      def put_header(name, value) = headers[name] = value
 
       # This method is used to set the request params.
       #
@@ -16,6 +33,7 @@ module Lenna
       # @return [Hash] the request params
       #
       # @api public
+      #
       def assign_params(params) = @params = params
 
       # This method is used to parse the body params.
@@ -23,6 +41,7 @@ module Lenna
       # @return [Hash] the request params
       #
       # @api public
+      #
       def params = super.merge(parse_body_params)
 
       # This method rewinds the body
@@ -32,9 +51,10 @@ module Lenna
       # @api public
       #
       # @since 0.1.0
-      def body_content
-        body.rewind
-        body.read
+      #
+      def body
+        super.rewind
+        super.read
       end
 
       # This method returns the headers in a normalized way.
@@ -46,6 +66,7 @@ module Lenna
       # @example:
       # Turn this:
       # HTTP_FOO=bar Foo=bar
+      #
       def headers
         content_type = env['CONTENT_TYPE']
         @headers ||= env.select { |k, _| k.start_with?('HTTP_') }
@@ -55,19 +76,30 @@ module Lenna
         @headers
       end
 
-      # This method returns the request body in a normalized way.
+      # This method returns the request body like a json.
       #
       # @return [Hash] the request body
       #
       # @api public
+      #
       def json_body = @json_body ||= parse_body_params
 
       private
 
-      def json_request? = media_type == 'application/json'
+      # This method returns the media type.
+      #
+      # @return [String] the request media type
+      #
+      def media_type = headers['Content-Type']
 
+      # This method parses the json body.
+      #
+      # @return [Hash] the request json body
+      #
+      # @api private
+      #
       def parse_json_body
-        @parsed_json_body ||= ::JSON.parse(body_content) if json_request?
+        @parsed_json_body ||= ::JSON.parse(body)
       rescue ::JSON::ParserError
         {}
       end
@@ -79,12 +111,8 @@ module Lenna
       # @api private
       def parse_body_params
         case media_type
-        when 'application/json'
-          parse_json_body
-        when 'application/x-www-form-urlencoded', 'multipart/form-data'
-          post_params
-        else
-          {}
+        in 'application/json' then parse_json_body
+        else post_params
         end
       end
 
@@ -93,12 +121,13 @@ module Lenna
       # @return [Hash] the request post params
       #
       # @api private
+      #
       def post_params
         @post_params ||=
-          if body_content.empty?
+          if body.empty?
             {}
           else
-            ::Rack::Utils.parse_nested_query(body_content)
+            ::Rack::Utils.parse_nested_query(body)
           end
       end
 
@@ -109,6 +138,7 @@ module Lenna
       # @return [String] the formatted header name
       #
       # @api private
+      #
       def format_header_name(name)
         name.sub(/^HTTP_/, '').split('_').map(&:capitalize).join('-')
       end