+++ lib/lenna/router/response.rb
@@ -9,24 +9,124 @@ module Lenna
     # @attr status  [Integer]       the response status
     # @attr params  [Hash]          the response params
     class Response
-      private attr_writer :body, :headers, :status, :params
-      public  attr_reader :body, :headers, :status, :params
+      # The status of the response.
+      #
+      # @return [Integer] the status of the Response
+      #
+      # @api public
+      #
+      # @example:
+      #   response.status = 200
+      #   # => 200
+      #
+      #   response.status
+      #   # => 200
+      #
+      public attr_accessor :status
 
+      # The body of the response.
+      #
+      # @return [Array(String)] the body of the Response
+      #
+      # @api public
+      #
+      # @example:
+      #
+      #   response.body
+      #   # => ['Hello, World!']
+      #
+      public attr_reader :body
+
+      # The headers of the response.
+      #
+      # @return [Hash] the headers of the Response
+      #
+      # @api public
+      #
+      # @example:
+      #
+      #   response.headers
+      #   # => { 'Content-Type' => 'text/html' }
+      #
+      public attr_reader :headers
+
+      # The params of the response.
+      #
+      # @return [Hash] the params of the Response
+      #
+      # @api public
+      #
+      # @example:
+      #
+      #   response.params
+      #   # => { 'name' => 'John' }
+      #
+      public attr_reader :params
+
+      # The length of the response.
+      #
+      # @return [Integer] the length of the Response
+      #
+      # @api public
+      #
+      # @example:
+      #
+      #   response.length
+      #   # => 0
+      #
+      public attr_reader :length
+
+      # This method will initialize the response.
+      #
+      # @param headers [Hash]          the response headers, default is {}
+      # @param status  [Integer]       the response status, default is 200
+      # @param body    [Array(String)] the response body, default is []
+      # @param params  [Hash]          the response params, default is {}
+      # @length        [Integer]      the response length, default is 0
+      #
+      # @return [void]
       def initialize(headers = {}, status = 200, body = [])
         @params  = {}
         @body    = body
         @status  = status
         @headers = headers
+        @length  = 0
       end
 
-      # This method will set the response status.
+      # This method will set the params.
       #
-      # @param value [Integer] the response status
+      # @param params [Hash] the params to be used
       #
-      # @return      [void]
+      # @return       [void]
       #
       # @api public
-      def assign_status(value) = put_status(value)
+      #
+      # @example:
+      #   response.params = { 'name' => 'John' }
+      #   # => { 'name' => 'John' }
+      #
+      def assign_params(params)
+        params => ::Hash
+
+        @params = params
+      rescue ::NoMatchingPatternError
+        raise ::ArgumentError, 'params must be a hash'
+      end
+      alias :params= assign_params
+
+      # Returns the response header corresponding to `key`.
+      #
+      # @param key [String] the header name
+      #
+      # @return    [String] the header value
+      #
+      # @api public
+      #
+      # @example:
+      #     res["Content-Type"]   # => "text/html"
+      #     res["Content-Length"] # => "42"
+      #
+      def [](key) = @headers[key]
 
       # Thi method set the body value.
       #
@@ -35,7 +135,9 @@ module Lenna
       # @return      [void]
       #
       # @api public
+      #
       def assign_body(value) = put_body(value)
+      alias :body= assign_body
 
       # This method will set the header value.
       #
@@ -57,7 +159,9 @@ module Lenna
       #
       #   assign_header('X-Request-Id', ['456', '789'])
       #   # => ['123', '456', '789']
+      #
       def assign_header(key, value) = put_header(key, value)
+      alias []= assign_header
 
       # Add multiple headers.
       #
@@ -78,12 +182,14 @@ module Lenna
 
         headers.each { |key, value| put_header(key, value) }
       end
+      alias :headers= assign_headers
 
       # This method will get the content type.
       #
       # @return [String] the content type
       #
       # @api public
+      #
       def content_type = @headers['Content-Type']
 
       # This method will delete the header.
@@ -92,23 +198,8 @@ module Lenna
       # @return       [void]
       #
       # @api public
+      #
       def remove_header(key) = delete_header(key)
-
-      # This method will get the redirect location.
-      #
-      # @param value [String] the key of the cookie
-      #
-      # @return      [String] the cookie
-      #
-      # @api public
-      def cookie(value)
-        value => ::String
-
-        @headers['Set-Cookie']
-          .then { |cookie| cookie.split('; ') }
-          .then { |cookie| cookie.find { |c| c.start_with?("#{value}=") } }
-          .then { |cookie| cookie.split('=').last }
-      end
 
       # This method will set the cookie.
       #
@@ -118,27 +209,70 @@ module Lenna
       # @return      [void]
       #
       # @api public
+      #
+      # @example:
+      #   response.assign_cookie('foo', 'bar')
+      #   # => 'foo=bar'
+      #
+      #   response.header['Set-Cookie']
+      #   # => 'foo=bar'
+      #
+      #  response.assign_cookie('foo2', 'bar2')
+      #  # => 'foo=bar; foo2=bar2'
+      #  response.header['Set-Cookie']
+      #  # => 'foo=bar; foo2=bar2'
+      #
+      #  response.assign_cookie('bar', {
+      #    domain: 'example.com',
+      #    path: '/',
+      #    # expires: Time.now + 24 * 60 * 60,
+      #    secure: true,
+      #    httponly: true
+      #  })
+      #
+      # response.header['Set-Cookie'].split('\n').last
+      # # => 'bar=; domain=example.com; path=/; secure; HttpOnly'
+      #
+      # @note:
+      #  This method doesn't sign and/or encrypt the cookie.
+      #  If you want to sign and/or encrypt the cookie, then you can use
+      #  the `Rack::Session::Cookie` middleware.
+      #
       def assign_cookie(key, value)
         key   => ::String
         value => ::String
 
-        cookie = "#{key}=#{value}"
-
-        put_header('Set-Cookie', cookie)
+        ::Rack::Utils.set_cookie_header!(@headers, key, value)
+      rescue ::NoMatchingPatternError
+        raise ::ArgumentError, 'key must be a string'
       end
 
       # This method will get all the cookies.
       #
       # @return [Hash] the cookies
+      #
       # @api public
-      def cookies
-        @headers['Set-Cookie']
-          .then { |cookie| cookie.split('; ') }
-          .each_with_object({}) do |cookie, acc|
-          key, value = cookie.split('=')
+      #
+      def cookies = @headers['set-cookie']
 
-          acc[key] = value
-        end
+      # This method will delete the cookie.
+      #
+      # @param key [String] the key of the cookie
+      #
+      # @return    [void]
+      #
+      # @api public
+      #
+      # @example:
+      #   response.delete_cookie('foo')
+      #   # => 'foo=; max-age=0; expires=Thu, 01 Jan 1970 00:00:00 -0000'
+      #
+      def delete_cookie(key)
+        key => ::String
+
+        ::Rack::Utils.delete_cookie_header!(@headers, key)
+      rescue ::NoMatchingPatternError
+        raise ::ArgumentError, 'key must be a string'
       end
 
       # This method will set redirect location. The status will be set to 302.
@@ -149,6 +283,7 @@ module Lenna
       # @return         [void]
       #
       # @api public
+      #
       def redirect(location, status: 302)
         location => ::String
 
@@ -165,6 +300,7 @@ module Lenna
       # @return [void]
       #
       # @api public
+      #
       def finish = finish!
 
       # This method will set the response content type.
@@ -175,6 +311,7 @@ module Lenna
       # @return        [void]
       #
       # @api public
+      #
       def assign_content_type(type, charset: nil)
         type => ::String
 
@@ -196,7 +333,15 @@ module Lenna
       # @return     [void]
       #
       # @api public
-      def json(data:, status: 200)
+      #
+      # @example:
+      #   response.json({ foo: 'bar' })
+      #   # => { foo: 'bar' }
+      #
+      #  response.json([{ foo: 'bar' }])
+      #  # => [{ foo: 'bar' }]
+      #
+      def json(data = {}, status: 200)
         data => ::Array | ::Hash
 
         put_status(status)
@@ -213,6 +358,7 @@ module Lenna
       # @return    [void]
       #
       # @api public
+      #
       def html(str = nil, status: 200)
         put_status(status)
         put_header('Content-Type', 'text/html')
@@ -242,6 +388,7 @@ module Lenna
       #   render('index', locals: { name: 'John' })
       #   # => Render the template `views/index.html.erb` with the local
       #   #    variable `name` set to 'John'
+      #
       def render(template_name, path: 'views', locals: {}, status: 200)
         template_path = ::File.join(path, "#{template_name}.html.erb")
 
@@ -273,6 +420,7 @@ module Lenna
       #
       # @see #render
       # @see #finish!
+      #
       def not_found
         put_body(['Not Found'])
         put_status(404)
@@ -287,6 +435,7 @@ module Lenna
       # @return [Integer] the response status
       #
       # @api private
+      #
       def put_status(value)
         value => ::Integer
 
@@ -298,7 +447,9 @@ module Lenna
       # This method will set the body.
       #
       # @param body [Array(String)] the body to be used
+      #
       # @return     [void]
+      #
       def put_body(value)
         value => ::String | ::Array
 
@@ -314,6 +465,7 @@ module Lenna
       # @param value [String] the value to be used
       #
       # @return      [void]
+      #
       def put_header(key, value)
         raise ::ArgumentError, 'key must be a string' unless key.is_a?(::String)
 
@@ -332,11 +484,13 @@ module Lenna
       # @param key [String] the header name
       #
       # @return    [void]
+      #
       def delete_header(key) = @headers.delete(key)
 
       # @param value [String] the redirect location
       #
       # @return      [void]
+      #
       def location!(value)
         value => ::String
 
@@ -346,6 +500,7 @@ module Lenna
       # This method will finish the response.
       #
       # @return [void]
+      #
       def finish!
         default_router_header!
         default_content_length!    unless @headers['Content-Length']
@@ -357,6 +512,7 @@ module Lenna
       # This method will set the response default html content type.
       #
       # @return [void]
+      #
       def default_html_content_type!
         put_header('Content-Type', 'text/html')
       end
@@ -364,6 +520,7 @@ module Lenna
       # This method will set the response default content length.
       #
       # @return [void]
+      #
       def default_content_length!
         put_header('Content-Length', @body.join.size.to_s)
       end
@@ -371,6 +528,7 @@ module Lenna
       # This method will set the response default router header.
       #
       # @return [void]
+      #
       def default_router_header!
         put_header('Server', "Lennarb VERSION #{::Lennarb::VERSION}")
       end