+++ test/lib/lenna/middleware/test_app.rb
@@ -5,24 +5,42 @@ require 'test_helper'
 module Lenna
   module Middleware
     class TestApp < Minitest::Test
-      def test_use
-        manager = App.new
-        mdw = ->(_req, _res, next_middleware) { next_middleware.call }
-        manager.use(mdw)
+      def setup
+        @app = App.instance
+      end
 
-        assert_equal manager.global_middlewares, [mdw]
+      def test_reset
+        mdw =
+          lambda do |_req, res, next_middleware|
+            res.assign_header('Custom-Header', 'From mdw')
+            next_middleware.call
+          end
+
+        @app.use(mdw)
+
+        assert_equal @app.global_middlewares, [mdw]
+      end
+
+      def test_use
+        mdw =
+          lambda do |_req, res, next_middleware|
+            res.assign_header('Custom-Header', 'From mdw')
+            next_middleware.call
+          end
+
+        @app.use(mdw)
+
+        assert_equal @app.global_middlewares, [mdw]
       end
 
       def test_fetch_or_build_middleware_chain
-        manager = App.new
-        action = ->(_req, _res) { 'Hello, World!' }
-        middleware_chain = manager.fetch_or_build_middleware_chain(action, [])
+        action = ->(_req, _res, next_middleware) { next_middleware.call }
+        middleware_chain = @app.fetch_or_build_middleware_chain(action, [])
 
         assert_equal middleware_chain, action
       end
 
       def test_middleware_execution_order_and_final_state
-        manager  = App.new
         env      = ::Rack::MockRequest.env_for('/')
         request  = Lenna::Router::Request.new(env)
         response = Lenna::Router::Response.new
@@ -38,9 +56,9 @@ module Lenna
             next_middleware.call
           end
 
-        manager.use([mdw1, mdw2])
+        @app.use([mdw1, mdw2])
         action = ->(_req, res) { [200, res.headers, ['OK']] }
-        chain = manager.build_middleware_chain(action, [])
+        chain = @app.build_middleware_chain(action, [])
         status, headers, body = chain.call(request, response)
 
         assert_equal(200, status)